+++ decidim-proposals/lib/decidim/proposals/admin_engine.rb
@@ -10,6 +10,9 @@ module Decidim
 
       routes do
         resources :proposals, only: [:index, :new, :create] do
+          collection do
+            resource :proposals_copy, only: [:new, :create]
+          end
           resources :proposal_answers, only: [:edit, :update]
           resources :proposal_notes, only: [:index, :create]
         end