+++ decidim-proposals/spec/system/admin_manages_proposals_spec.rb
@@ -17,4 +17,5 @@ describe "Admin manages proposals", type: :system do
   it_behaves_like "export proposals"
   it_behaves_like "manage announcements"
   it_behaves_like "manage proposals help texts"
+  it_behaves_like "copy proposals"
 end